made replication processor nicer to look at and added change detection event for blocks. also made a basic health pogress bar that reacts to that health change

+++ Main/Server/Processors/HealthProcessor.cs
@@ -32,13 +32,13 @@ public class HealthProcessor : Processor
         IncreaseHealth(entity, healthBlock.HealthRegen);
     }
 
-    public void IncreaseHealth(Entity entity, int change)
+    public void IncreaseHealth(Entity entity, float change)
     {
         var healthBlock = entity.GetBlock<Blocks.HealthBlock>();
         healthBlock.Health = Math.Min(healthBlock.MaxHealth, healthBlock.Health + change);
     }
 
-    public void DecreaseHealth(Entity entity, int change)
+    public void DecreaseHealth(Entity entity, float change)
     {
         var healthBlock = entity.GetBlock<Blocks.HealthBlock>();
         healthBlock.Health = Math.Max(0, healthBlock.Health - change);
